use int32 instead of int

+++ pkg/convert/convert.go
@@ -58,29 +58,29 @@ func Int64E(raw interface{}) (int64, error) {
 	}
 }
 
-// Int converts anything into a int
+// Int converts anything into a int32
 // errors will fall back to 0
-func Int(raw interface{}) int {
+func Int(raw interface{}) int32 {
 	val, _ := IntE(raw)
 
 	return val
 }
 
-// IntE converts anything into a int
+// IntE converts anything into a int32
 // errors will be returned
-func IntE(raw interface{}) (int, error) {
+func IntE(raw interface{}) (int32, error) {
 	switch val := raw.(type) {
 	case int:
-		return val, nil
+		return int32(val), nil
 	case int32:
-		return int(val), nil
+		return val, nil
 	default:
 		num, err := strconv.ParseInt(fmt.Sprintf("%v", val), 10, 32)
 		if err != nil {
 			return 0, fmt.Errorf("cannot parse int64 value from %v (%T)", raw, raw)
 		}
 
-		return int(num), nil
+		return int32(num), nil
 	}
 }
